tests/wpt/css-tests/css-transforms-1_dev/html/transform_translate_invalid.htm

<!DOCTYPE html>
<html><head>
    <title>CSS Transforms API Test: transform translate with invalid translation value</title>
    <link href="http://www.intel.com" rel="author" title="Intel">
    <link href="http://www.w3.org/TR/css-transforms-1/#transform-property" rel="help">
    <link href="http://www.w3.org/TR/css-transforms-1/#two-d-transform-functions" rel="help">
    <meta content="dom" name="flags">
    <meta content="Check if transform sets translate(null, null) that an expection is to be thrown" name="assert"> 
    <script src="/resources/testharness.js" type="text/javascript"></script>
    <script src="/resources/testharnessreport.js" type="text/javascript"></script>    
</head>
<body>
    <div id="test"></div>
    <div id="log"></div>
    <script type="text/javascript">
    test(function() {
        document.getElementById("test").style.transform = "translate(null, null)";
 
        assert_throws(null, function() {
            document.getElementById("test").style.getPropertyCSSValue("transform").cssText;
        });
    }, "transform_translate_null_null");
    </script>    


</body></html>
